Moving Teams Api calls to bot
